chore: Move schema source/generation out of samtranslator package
