+++ MANIFEST.in
@@ -5,10 +5,6 @@ recursive-include samtranslator/validator/sam_schema *.json
 include samtranslator/policy_templates_data/policy_templates.json
 include samtranslator/policy_templates_data/schema.json
 include samtranslator/model/connector_profiles/profiles.json
-include samtranslator/schema/docs.json
-include samtranslator/schema/schema.json
-include samtranslator/schema/sam.schema.json
-include samtranslator/schema/cloudformation.schema.json
 include README.md
 include THIRD_PARTY_LICENSES
