+++ requirements/base.txt
@@ -1,4 +1,3 @@
 boto3>=1.19.5,==1.*
 jsonschema<5,>=3.2  # TODO: evaluate risk of removing jsonschema 3.x support
-pydantic~=1.8
 typing_extensions~=4.4.0 # 3.7 doesn't have Literal
