+++ samtranslator/model/apigateway.py
@@ -5,8 +5,7 @@ from typing import Any, Dict, List, Optional
 from samtranslator.model import PropertyType, Resource
 from samtranslator.model.exceptions import InvalidResourceException
 from samtranslator.model.intrinsics import fnSub, ref
-from samtranslator.model.types import IS_DICT, IS_STR, is_type, list_of, one_of
-from samtranslator.schema.common import PassThrough
+from samtranslator.model.types import IS_DICT, IS_STR, is_type, list_of, one_of, PassThrough
 from samtranslator.translator import logical_id_generator
 from samtranslator.translator.arn_generator import ArnGenerator
 from samtranslator.utils.py27hash_fix import Py27Dict, Py27UniStr
