+++ schema_source/aws_serverless_function.py
@@ -4,7 +4,7 @@ from typing import Optional, Dict, Union, List
 
 from typing_extensions import Literal
 
-from samtranslator.schema.common import (
+from schema_source.common import (
     PassThroughProp,
     BaseModel,
     SamIntrinsicable,
@@ -13,7 +13,7 @@ from samtranslator.schema.common import (
     Ref,
     ResourceAttributes,
 )
-from samtranslator.schema.aws_serverless_connector import EmbeddedConnector
+from schema_source.aws_serverless_connector import EmbeddedConnector
 
 
 alexaskilleventproperties = get_prop("sam-property-function-alexaskill")
